Play around with shuffling things around (WIP)
